Add Theseus Cargo/Transport Vans
